🧪 YP9QCH task: keep context tests within hotspot budget
