+++ .agentplane/tasks/202607221852-YP9QCH/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202607221852-YP9QCH/build-source-driven-canonical-reconciliation-can",
   "created_at": "2026-07-30T09:44:33.829Z",
-  "diffstat_sha256": "sha256:46563ffa6cab4e77771240466c47e0e723b12bf087065083b13d50849aa93fac",
+  "diffstat_sha256": "sha256:f6a8e6de2ff671447923922a4b5831c332b908e355d39c16a572dfce2c8a07c2",
   "last_verified_at": "2026-07-30T10:19:14.880Z",
   "last_verified_diffstat_sha256": "sha256:46563ffa6cab4e77771240466c47e0e723b12bf087065083b13d50849aa93fac",
   "pr_number": 4688,
